+++ src/Repositories/CategoryRepositoryInterface.php
@@ -1,0 +1,45 @@
+<?php
+
+/**
+ * TechDivision\Import\Category\Ee\Repositories\CategoryRepositoryInterface
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-category-ee
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Category\Ee\Repositories;
+
+use TechDivision\Import\Repositories\RepositoryInterface;
+
+/**
+ * Interface for repository implementations to load category data for Magento 2 EE.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-category-ee
+ * @link      http://www.techdivision.com
+ */
+interface CategoryRepositoryInterface extends \TechDivision\Import\Category\Repositories\CategoryRepositoryInterface
+{
+
+    /**
+     * Return's the category with the passed ID.
+     *
+     * @param string $id The ID of the category to return
+     *
+     * @return array The category
+     */
+    public function load($id);
+}